% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_impute_AR1_Gaussian.R
\name{estimateAR1Gaussian}
\alias{estimateAR1Gaussian}
\title{Fit Gaussian AR(1) Model to Time Series with Missing Values}
\usage{
estimateAR1Gaussian(y, random_walk = FALSE, zero_mean = TRUE,
  iterates = FALSE, condMeanCov = FALSE, tol = 1e-10,
  maxiter = 1000)
}
\arguments{
\item{y}{a xts object indicating time series with missing values. The first and last one should not be NA.}

\item{random_walk}{logical. If TRUE, y is a random walk time series, and phi1 = 1. If FALSE, y is a general AR(1) time series, and phi1 is unknown. The default value is FALSE.}

\item{zero_mean}{logical. If TRUE, y is a zero-mean time series, and phi0 = 1. If FALSE, y is a general AR(1) time series, and phi0 is unknown. The default value is FALSE.}

\item{iterates}{logical. If TRUE, then the iterates are outputted. If FALSE, they are ignored. The default value is FALSE.}

\item{maxiter}{a positive integer indicating the maximum number of iterations allowed (default \code{1000}).}

\item{ftol}{a positive number controlling the stopping criterion (default \code{1e-8}).}
}
\value{
A list containing the following elements:
\item{\code{phi0}}{real number, the estimate for phi0}
\item{\code{phi1}}{real number, the estimate for phi1}
\item{\code{sigma2}}{positive number, the estimate for sigma^2}
\item{\code{phi0_iterate}}{a numerical vector, the estimates for phi0 in each iteration}
\item{\code{phi1_iterate}}{a numerical vector, the estimates for phi1 in each iteration}
\item{\code{sigma2_iterate}}{a vector of positive numbers, the estimates for sigma^2 in each iteration}
\item{\code{f_iterate}}{a numerical vector, the objective values in each iteration}
}
\description{
Estimate the parameters of a Gaussian AR(1) model from a time series with missing values
}
\examples{
library(imputeFin)
library(xts)
# generate a complete Student's t AR(1) time series
phi0 <- 0
phi1 <- 1
sigma2 <- 0.01 
nu <- 1
n <- 200
n_miss <- 25 
n_drop <- 100
n_total <- n + n_drop
data <- vector(length = n_total)
epsilon <- vector(length = n_total - 1)# innovations
data[1] <- 0
for (i in 2:n_total) {
  epsilon[i-1] <- rnorm(1, nu) * sqrt(sigma2)
  data[i] <- phi0 + phi1 * data[i-1] + epsilon[i-1]
}
data <- data[(n_drop + 1):n_total] # drop the first n_drop to reduce the influence of initial point
dates <- seq(as.Date("2016-01-01"), length = n, by = "days") 
y_orig <- xts(data,  dates)

# creat missing values
index_miss <- sample( 2:(n - 1), n_miss, FALSE)
index_miss <- sort(index_miss)
y <- y_orig
y[index_miss] <- NA

# estimate the parameters from this incomplete time series
estimation_result <- estimateAR1Gaussian(y)

}
\references{
R. J. Little and D. B. Rubin, Statistical Analysis with Missing Data, 2nd ed. Hoboken, N.J.: John Wiley & Sons, 2002.
}
\author{
Junyan Liu and Daniel P. Palomar
}
