% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_impute_AR1_Gaussian.R
\name{imputeAR1Gaussian}
\alias{imputeAR1Gaussian}
\title{Imputate Missing Values in  Incomplete Gaussian AR(1) Time Series}
\usage{
imputeAR1Gaussian(y, n_samples = 1, random_walk = FALSE,
  zero_mean = TRUE, estimates = FALSE, positions_NA = FALSE)
}
\arguments{
\item{y}{a xts object indicating time series with missing values. The first and last one should not be NA.}

\item{n_samples}{a positive integer indicating the number of imputations (default \code{1}).}

\item{random_walk}{logical. If TRUE, y is a random walk time series, and phi1 = 1. If FALSE, y is a general AR(1) time series, and phi1 is unknown. The default value is FALSE.}

\item{zero_mean}{logical. If TRUE, y is a zero-mean time series, and phi0 = 1. If FALSE, y is a general AR(1) time series, and phi0 is unknown.}

\item{param}{a list consisting of the paramters of the Student's t AR(1) time series y if known. The default value is FALSE.}
}
\value{
\item{\code{y_imputed}  }{a numerical matrix, each column is a imputed complete time series}
}
\description{
Estimate the parameters of the Gaussian AR(1) model from a time series with missing values and impute the missing values based on the estimates
}
\examples{
library(imputeFin)
library(xts)
phi0 <- 1
phi1 <- 0.5 
sigma2 <- 0.01 
nu <- 1
n <- 200
n_miss <- 25 
n_drop <- 100
n_total <- n + n_drop
data <- vector(length = n_total)
epsilon <- vector(length = n_total - 1)# innovations
data[1] <- 0
for (i in 2:n_total) {
  epsilon[i - 1] <- rnorm(1, nu) * sqrt(sigma2)
  data[i] <- phi0 + phi1 * data[i - 1] + epsilon[i - 1]
}
data <- data[(n_drop + 1):n_total] # drop the first n_drop to reduce the influence of initial point
dates <- seq(as.Date("2016-01-01"), length = n, by = "days") 
y_orig <- xts(data,  dates)

# creat missing values
index_miss <- sample(2:(n - 1), n_miss, FALSE)
index_miss <- sort(index_miss)
y <- y_orig
y[index_miss] <- NA

# impute the missing values and generate n_samples complete time series
y_imputed <- imputeAR1Gaussian( y_miss, n_samples = 3) # if the parameters are unknown
param = list("phi0" = phi0,
             "phi1" = phi1,
             "sigma2" = sigma2,
             "nu" = nu)
y_imputed <- imputeAR1Gaussian(y_miss, n_samples = 3, param) # if the parameters are unknown
}
\author{
Junyan Liu and Daniel P. Palomar
}
