% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_VAR_t.R
\docType{data}
\name{ts_VAR_t}
\alias{ts_VAR_t}
\title{Synthetic Student's t VAR data with missing values}
\format{
List with the following elements:
\describe{
  \item{Y}{200 x 3 \code{zoo} object as a Student's t VAR time series.}
  \item{phi0}{True value of the constant vector in the VAR model.}
  \item{Phii}{True value of the coefficient matrix in the VAR model.}
  \item{scatter}{True value of the scatter matrix (of the noise distribution) in the VAR model.}
  \item{nu}{True value of the degrees of freedom (of the noise distribution) in the VAR model.}
}
}
\usage{
data(ts_VAR_t)
}
\description{
Synthetic Student's t VAR data with missing values for 
estimation and imputation testing purposes.
}
\keyword{dataset}
